Move contexts to own package
